Fixed fireray and lightning ray logic
- Added new class LineCasting for linecasting
- Added damage ticks per second in fireray and lightning ray
