--- Assets/Prefabs/Spells/Fireray.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f4e8061d9e650b942bf22e546bbfb284
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
