+++ Assets/Prefabs/Spells/Fireray.prefab
@@ -23841,7 +23841,7 @@ GameObject:
   - component: {fileID: 5149430543365902812}
   - component: {fileID: 7729258316315907497}
   m_Layer: 8
-  m_Name: Laser
+  m_Name: Fireray
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -23877,7 +23877,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b1c35bef718746e4fb026a1efd632768, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damagePerFrame: 5
+  damage: 10
+  damageTicksPerSecond: 8
 --- !u!65 &7729258316315907497
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -23889,5 +23890,5 @@ BoxCollider:
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 4, y: 5, z: 1}
-  m_Center: {x: 0, y: 0, z: 0.5}
+  m_Size: {x: 4, y: 5, z: 0.88}
+  m_Center: {x: 0, y: 0, z: 0.45}
