--- Assets/Prefabs/Spells/Fireray/Fireray.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 95c382ab29425d9429f77a5021855fe4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
