+++ Assets/Scripts/Spells/Fireray.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b183d8ecabf42904ebbcfb6963aa1302
+guid: b1c35bef718746e4fb026a1efd632768
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
